refactor: 기존 category custom repository를 분리하고 재구성 - #187
DB와 관련 없는 작업을 하는 메서드는 util 함수로 분리하고,
관련 있는 메서드를 다시 repository로 다시 작성까지 완료
