+++ src/collections/collections.service.ts
@@ -22,7 +22,7 @@ import {
 } from './dtos/nested-content.dto';
 import { toggleFavoriteOutput } from 'src/contents/dtos/content.dto';
 import { Category } from '../contents/entities/category.entity';
-import { CategoryRepository } from '../contents/repository/category.repository';
+import { CategoryRepository } from '../contents/repository/category.old.repository';
 import { ContentsService } from '../contents/contents.service';
 
 @Injectable()
